refactor: move h3 augmentation to src/types/h3
