+++ src/types/h3.ts
@@ -1,0 +1,15 @@
+import type { NitroFetchRequest, $Fetch } from "./fetch";
+
+declare module "h3" {
+  interface H3Event {
+    /** @experimental Calls fetch with same context and request headers */
+    fetch: (
+      request: NitroFetchRequest,
+      init?: RequestInit
+    ) => Promise<Response>;
+    /** @experimental Calls fetch with same context and request headers */
+    $fetch: $Fetch<unknown, NitroFetchRequest>;
+  }
+}
+
+export {};
