+++ src/types/index.ts
@@ -1,5 +1,6 @@
 export * from "./fetch";
 export * from "./global";
+export * from "./h3";
 export * from "./serialize";
 export * from "./nitro";
 export * from "./handler";
