+++ types.d.ts
@@ -1,2 +1,4 @@
-export * from './dist/index'
-export * from './dist/runtime/types'
+export * from "./dist/index";
+export * from "./dist/runtime/types";
+
+import type { NitroFetchRequest, $Fetch } from "./dist/index";
